requirements.txt
autopep8==1.5.2
click==7.1.2
Flask==1.1.2
Flask-SocketIO==4.3.0
gevent==20.6.0
greenlet==0.4.16
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
python-dotenv==0.13.0
python-engineio==3.13.0
python-socketio==4.6.0
six==1.15.0
stringcase==1.2.0
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
